encapsulate Checkpoint internal state
